requirements.txt
--find-links https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio

tensorboardX
tensorboard
tensorflow
easydict
pandas
numpy
scipy
opencv-python
pyyaml
yacs

# new installments
matplotlib
tqdm
timm
ipykernel
scikit-image
scikit-learn
cv2
pillow